+++ src/components/settings/setting-container.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+    export let dataname: string;
+    export let settingItems;
+</script>
+
+<div class="config__tab-container" data-name={dataname}>
+
+</div>
