+++ src/components/settings/setting-item.svelte
@@ -1,6 +1,6 @@
 <script>
     import { createEventDispatcher } from "svelte";
-    import { eventBus } from "../event-bus";
+    import { eventBus } from "@/event-bus";
     export let type; // 设置项目类型
     export let content; // 设置项目内部文本展示
     export let settingKey; // 设置项目 key
